Feat: Ticket entity and refactored this code for join Order, OrderItem, TicketType and Ticket

+++ backend/src/main/java/io/github/fabiocintra/event_management/order/model/dto/OrderRequest.java
@@ -9,6 +9,6 @@ public record OrderRequest(
         @NotBlank(message = "Campo obrigatorio!")
         String attendeeId,
         @NotNull(message = "Campo obrigatorio!")
-        BigDecimal totalAmount
+        Double totalAmount
 ) {
 }
